Move Dtos and common files in Contracts project in order to avoid ciricular dependencies between projects.
