--- ParkBuddy.Application/ApiResponse.cs
@@ -1,16 +1,0 @@
-namespace ParkBuddy.Application
-{
-    public class ApiResponse<T>
-    {
-        public bool Success { get; set; }
-        public string Message { get; set; }
-        public T? Data { get; set; }
-
-        public ApiResponse(bool success, string message, T? data = default) // default value in null
-        {
-            Success = success;
-            Message = message;
-            Data = data;
-        }
-    }
-}
